breaking: IOconfFileLoader is no longer public
By not having this public, we can change some of the underlying details without further breaking changes.
Use IOconfFile.DefaultLoader and IOconfFile.DefaultFileExists instead.

Also fixed redundancy support in the data uploader host due to it being added to late in the program initialization
