+++ CA_DataUploader/Program.cs
@@ -18,11 +18,12 @@ namespace CA_DataUploader
         {
             try
             {
-                var loglevel = IOconfFileLoader.FileExists() ? IOconfFile.Instance.GetOutputLevel() : IOconfLoopName.Default.LogLevel;
+                Redundancy.RegisterSystemExtensions(IOconfFile.DefaultLoader);
+                var conf = IOconfFile.FileExists() ? IOconfFile.Instance : null;
+                var loglevel = conf?.GetOutputLevel() ?? IOconfLoopName.Default.LogLevel;
                 CALog.LogInfoAndConsoleLn(LogID.A, RpiVersion.GetWelcomeMessage($"Upload temperature data to cloud", loglevel));
                 Console.WriteLine("Initializing...");
-                Redundancy.RegisterSystemExtensions(IOconfFileLoader.Loader);
-                var serial = new SerialNumberMapper(IOconfFileLoader.FileExists() ? IOconfFile.Instance : null);
+                var serial = new SerialNumberMapper(conf);
                 await serial.DetectDevices();
 
                 if (args.Length > 0 && args[0] == "-listdevices")
